requirements.txt
fastapi
uvicorn
langchain
PyPDF2
chromadb
tiktoken
google-generativeai